% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway.R
\name{Pathway}
\alias{Pathway}
\title{Pathway}
\value{
a named list, indexed by species ids, with reactome data in each element
}
\description{
an R6 class which parses and represents a Reactome Pathway
}
\details{
import R6
import xml2
import data.table
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Pathway$new()}}
\item \href{#method-getName}{\code{Pathway$getName()}}
\item \href{#method-getReactionCount}{\code{Pathway$getReactionCount()}}
\item \href{#method-getReactionNames}{\code{Pathway$getReactionNames()}}
\item \href{#method-getPathwayNameMap}{\code{Pathway$getPathwayNameMap()}}
\item \href{#method-getMolecularSpeciesMap}{\code{Pathway$getMolecularSpeciesMap()}}
\item \href{#method-clone}{\code{Pathway$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Pathway object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pathway$new(sbml.filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sbml.filename}}{character string, path to a valid sbml 3.1 document}

\item{\code{xml.node}}{an XMLInternalNode}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ReactionParser` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
every Reactome pathway has a name, e.g., R-HSA-165159 is "MTOR signalling"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pathway$getName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character string, the pathway name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReactionCount"></a>}}
\if{latex}{\out{\hypertarget{method-getReactionCount}{}}}
\subsection{Method \code{getReactionCount()}}{
every Reactome pathway has multiple reactions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pathway$getReactionCount()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
numeric count
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReactionNames"></a>}}
\if{latex}{\out{\hypertarget{method-getReactionNames}{}}}
\subsection{Method \code{getReactionNames()}}{
every Reactome pathway has multiple reactions, each with a name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pathway$getReactionNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPathwayNameMap"></a>}}
\if{latex}{\out{\hypertarget{method-getPathwayNameMap}{}}}
\subsection{Method \code{getPathwayNameMap()}}{
many pathways
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pathway$getPathwayNameMap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the map
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMolecularSpeciesMap"></a>}}
\if{latex}{\out{\hypertarget{method-getMolecularSpeciesMap}{}}}
\subsection{Method \code{getMolecularSpeciesMap()}}{
molecular species from xml hierarchy as list list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pathway$getMolecularSpeciesMap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pathway$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
