% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReactionParser.R
\name{ReactionParser}
\alias{ReactionParser}
\title{ReactionParser}
\value{
an XMLInternalNode

a character string, e.g, "/sbml/model/listOfReactions/reaction[5]"

an XMLInternalNode

a character string, the id for this reaction node

a character string, the name of this reaction node

a character string, the compartment in which  this reaction takes place

a character string, the short-to-mediums account of the reaction

an integer count

reactant (species) ids

an integer count

product (species) ids

an integer count

character (species) ids

an integer count

an named list

a named list, complex name and complex participatns

one of "drug", "reaction", "protein", "ligand", or "unrecognized"

one of "drug", "reaction", "protein", "ligand", or "unrecognized"

a named list, edges and nodes, each a data.frame
}
\description{
an R6 class which parses xml Reactome reactions into data.frames
}
\details{
import R6
import xml2
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReactionParser$new()}}
\item \href{#method-getXPath}{\code{ReactionParser$getXPath()}}
\item \href{#method-getXml}{\code{ReactionParser$getXml()}}
\item \href{#method-getID}{\code{ReactionParser$getID()}}
\item \href{#method-getName}{\code{ReactionParser$getName()}}
\item \href{#method-getCompartment}{\code{ReactionParser$getCompartment()}}
\item \href{#method-getNotes}{\code{ReactionParser$getNotes()}}
\item \href{#method-getReactantCount}{\code{ReactionParser$getReactantCount()}}
\item \href{#method-getReactants}{\code{ReactionParser$getReactants()}}
\item \href{#method-getProductCount}{\code{ReactionParser$getProductCount()}}
\item \href{#method-getProducts}{\code{ReactionParser$getProducts()}}
\item \href{#method-getModifierCount}{\code{ReactionParser$getModifierCount()}}
\item \href{#method-getModifiers}{\code{ReactionParser$getModifiers()}}
\item \href{#method-getComplexCount}{\code{ReactionParser$getComplexCount()}}
\item \href{#method-getCounts}{\code{ReactionParser$getCounts()}}
\item \href{#method-getComplexes}{\code{ReactionParser$getComplexes()}}
\item \href{#method-assignNodeType}{\code{ReactionParser$assignNodeType()}}
\item \href{#method-assignNodeName}{\code{ReactionParser$assignNodeName()}}
\item \href{#method-toEdgeAndNodeTables}{\code{ReactionParser$toEdgeAndNodeTables()}}
\item \href{#method-clone}{\code{ReactionParser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new parser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$new(doc, xml.node, parent.pathway)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{an xml_document}

\item{\code{xml.node}}{an XMLInternalNode}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ReactionParser` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getXPath"></a>}}
\if{latex}{\out{\hypertarget{method-getXPath}{}}}
\subsection{Method \code{getXPath()}}{
easy access to the entire xml element


XPath from document root for the reaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getXPath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getXml"></a>}}
\if{latex}{\out{\hypertarget{method-getXml}{}}}
\subsection{Method \code{getXml()}}{
easy access to the entire xml element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getXml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\if{latex}{\out{\hypertarget{method-getID}{}}}
\subsection{Method \code{getID()}}{
every reaction has a reactome identifier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
every reaction has a name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCompartment"></a>}}
\if{latex}{\out{\hypertarget{method-getCompartment}{}}}
\subsection{Method \code{getCompartment()}}{
a reaction is specific to a cellular compartmentevery reaction has a name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getCompartment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNotes"></a>}}
\if{latex}{\out{\hypertarget{method-getNotes}{}}}
\subsection{Method \code{getNotes()}}{
every (most?) reactions are accompanied by explanatory notes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getNotes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReactantCount"></a>}}
\if{latex}{\out{\hypertarget{method-getReactantCount}{}}}
\subsection{Method \code{getReactantCount()}}{
one (zero?) or more reactants contribute to each reaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getReactantCount()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReactants"></a>}}
\if{latex}{\out{\hypertarget{method-getReactants}{}}}
\subsection{Method \code{getReactants()}}{
get the species identifiers for all reactants
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getReactants()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProductCount"></a>}}
\if{latex}{\out{\hypertarget{method-getProductCount}{}}}
\subsection{Method \code{getProductCount()}}{
one (zero?) or more products are produced by each reaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getProductCount()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProducts"></a>}}
\if{latex}{\out{\hypertarget{method-getProducts}{}}}
\subsection{Method \code{getProducts()}}{
get the species identifiers for all products
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getProducts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getModifierCount"></a>}}
\if{latex}{\out{\hypertarget{method-getModifierCount}{}}}
\subsection{Method \code{getModifierCount()}}{
zero or more modifiers are active in each reaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getModifierCount()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getModifiers"></a>}}
\if{latex}{\out{\hypertarget{method-getModifiers}{}}}
\subsection{Method \code{getModifiers()}}{
get the species identifiers for all modifiers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getModifiers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getComplexCount"></a>}}
\if{latex}{\out{\hypertarget{method-getComplexCount}{}}}
\subsection{Method \code{getComplexCount()}}{
zero or more molecular complexes participate in this reaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getComplexCount()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCounts"></a>}}
\if{latex}{\out{\hypertarget{method-getCounts}{}}}
\subsection{Method \code{getCounts()}}{
reactants, products, modifiers, complexes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getCounts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getComplexes"></a>}}
\if{latex}{\out{\hypertarget{method-getComplexes}{}}}
\subsection{Method \code{getComplexes()}}{
identify molecular species which are complexes of molecular species
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$getComplexes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignNodeType"></a>}}
\if{latex}{\out{\hypertarget{method-assignNodeType}{}}}
\subsection{Method \code{assignNodeType()}}{
reactome entities (nodes) are of identifiable types; assign them here
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$assignNodeType(node.id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node.id}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignNodeName"></a>}}
\if{latex}{\out{\hypertarget{method-assignNodeName}{}}}
\subsection{Method \code{assignNodeName()}}{
reactome entities (nodes) are of identifiable types; assign them here
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$assignNodeName(node.id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node.id}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toEdgeAndNodeTables"></a>}}
\if{latex}{\out{\hypertarget{method-toEdgeAndNodeTables}{}}}
\subsection{Method \code{toEdgeAndNodeTables()}}{
cytoscape.js various databases (sql, neo4j, dc) represent data in tables.
create them here
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$toEdgeAndNodeTables(
  excludeUbiquitousSpecies = TRUE,
  includeComplexMembers
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{excludeUbiquitousSpecies}}{logical, default TRUE: ATP, ADP, water}

\item{\code{includeComplexMembers}}{logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionParser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
